+++ tests/tests_utils/test_benchmark_gather_rosbag_trajectory_window_informations.py
@@ -5,7 +5,7 @@ from typing import Optional, Tuple, Union
 
 import pytest
 
-from tests.rosbag_test_utils import (
+from tests.testing_utils_rosbag import (
     get_rosbag_vaul_f1tenth_nx_orin_path_filtered_short,
     get_rosbag_vaul_f1tenth_nx_orin_path_offending_timestamps,
 )
